dwarf/op,proc: output register name when printing location exprs
Output the architecture specific register name when printing location
expression after DW_OP_regN and DW_OP_bregN operators.

+++ pkg/dwarf/regnum/i386.go
@@ -93,8 +93,8 @@ func I386MaxRegNum() int {
 	return max
 }
 
-func I386ToName(num int) string {
-	name, ok := i386DwarfToName[num]
+func I386ToName(num uint64) string {
+	name, ok := i386DwarfToName[int(num)]
 	if ok {
 		return name
 	}
